refactor(editor): move editor into web-pkg
