--- packages/editor/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@opencloud-eu/editor",
-  "version": "0.0.0",
-  "private": true,
-  "description": "Content-type-aware rich text editor built on Tiptap",
-  "license": "AGPL-3.0",
-  "main": "src/index.ts",
-  "type": "module",
-  "dependencies": {
-    "@tiptap/core": "^3.20.4",
-    "@tiptap/extension-document": "^3.20.4",
-    "@tiptap/extension-hard-break": "^3.20.4",
-    "@tiptap/extension-image": "^3.20.4",
-    "@tiptap/extension-link": "^3.20.4",
-    "@tiptap/extension-paragraph": "^3.20.4",
-    "@tiptap/extension-text": "^3.20.4",
-    "@tiptap/extension-table": "^3.20.4",
-    "@tiptap/extension-table-cell": "^3.20.4",
-    "@tiptap/extension-table-header": "^3.20.4",
-    "@tiptap/extension-table-row": "^3.20.4",
-    "@tiptap/extension-task-item": "^3.20.4",
-    "@tiptap/extension-task-list": "^3.20.4",
-    "@tiptap/extension-underline": "^3.20.4",
-    "@tiptap/markdown": "^3.20.4",
-    "@tiptap/pm": "^3.20.4",
-    "@tiptap/starter-kit": "^3.20.4",
-    "@tiptap/suggestion": "^3.20.4",
-    "@tiptap/vue-3": "^3.20.4"
-  },
-  "peerDependencies": {
-    "@opencloud-eu/design-system": "workspace:^",
-    "vue": "^3.5.0",
-    "vue3-gettext": "^4.0.0-beta.1"
-  },
-  "devDependencies": {
-    "@opencloud-eu/web-test-helpers": "workspace:*"
-  }
-}
