--- packages/editor/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@opencloud-eu/tsconfig"
-}
