+++ packages/web-app-app-store/package.json
@@ -9,7 +9,6 @@
   },
   "peerDependencies": {
     "@opencloud-eu/design-system": "workspace:^",
-    "@opencloud-eu/editor": "workspace:*",
     "@opencloud-eu/web-client": "workspace:*",
     "@opencloud-eu/web-pkg": "workspace:*",
     "axios": "1.15.2",
