+++ packages/web-app-app-store/src/views/AppDetails.vue
@@ -55,7 +55,7 @@ import { computed, defineComponent, unref, watch } from 'vue'
 import { App } from '../types'
 import { APPID } from '../appid'
 import { useRouteParam, useRouter } from '@opencloud-eu/web-pkg'
-import { useTextEditor, TextEditorProvider, TextEditorContent } from '@opencloud-eu/editor'
+import { useTextEditor, TextEditorProvider, TextEditorContent } from '@opencloud-eu/web-pkg'
 import { useAppsStore } from '../piniaStores'
 import AppResources from '../components/AppResources.vue'
 import AppTags from '../components/AppTags.vue'
