+++ packages/web-pkg/src/editor/index.ts
@@ -1,13 +1,6 @@
-export type {
-  ContentType,
-  TextEditorOptions,
-  TextEditorInstance,
-  SlashCommandItem,
-  SlashCommandGroup
-} from './types'
-export type { ToolbarItem, ToolbarGroup } from './actions/toolbar/types'
+export type { ContentType, TextEditorOptions, TextEditorInstance } from './types'
+export type { ToolbarItem, ToolbarGroup, SlashCommandItem, SlashCommandGroup } from './actions'
 export { useTextEditor } from './composables/useTextEditor'
 export { default as TextEditorProvider } from './components/TextEditorProvider.vue'
 export { default as TextEditorContent } from './components/TextEditorContent.vue'
 export { default as TextEditorToolbar } from './components/TextEditorToolbar.vue'
-export { richTextSlashCommandGroups } from './actions/slashCommands/items'
