+++ packages/web-pkg/src/index.ts
@@ -2,6 +2,7 @@ export * from './apps'
 export * from './components'
 export * from './composables'
 export * from './constants'
+export * from './editor'
 export * from './errors'
 export * from './extensionPoints'
 export * from './helpers'
